Fixed text issues
